progress: make -r reboot only if something was installed
swupdate-progress may be used as a convenient tool to monitor updates
happening in the background. It also provides an option to reboot the
system when the installation completes. SWUpdate may send a "SUCCESS"
status even if no artifacts were installed (possiby because all images
had install-if-higher and the current system already running the latest
and greatest version of the OS). Avoid a reboot by checking cur_step
from the status update messages to determine what to do on "SUCCESS".

+++ tools/swupdate-progress.c
@@ -324,7 +324,7 @@ int main(int argc, char **argv)
 			if (psplash_ok)
 				psplash_progress(psplash_pipe_path, &msg);
 			psplash_ok = 0;
-			if ((msg.status == SUCCESS) && opt_r) {
+			if ((msg.status == SUCCESS) && (msg.cur_step > 0) && opt_r) {
 				sleep(5);
 				if (system("reboot") < 0) { /* It should never happen */
 					fprintf(stdout, "Please reset the board.\n");
